DATASOLR-6 - Added XML namespace and JavaConfig support.
Introduced SolrRepositoryFactoryBean to ease setting up repository beans by namespace parsers and the annotation infrastructure. Added namespace XSD with <solr:repositories /> element and @EnableSolrRepositories as well as the necessary glue code to trigger repository scanning and bean creation.

In the integration test for the XML namespace we currently need to work around the issue that a SolrServer instance cannot be easily set up using XML configuration. Thus we manually register the instance created by the superclass with the ApplicationContext created for this test case. This should be removed once we have decent support for setting up an embedded SolrServer via Spring Config.