DATASOLR-57 - Upgrade to solr 4.2.0
Upgrade showed following glitches

- ParseException moved to org.apache.solr.parser.ParseException:
The import within the test case was updated. ExceptionTranslater happily
handles the error already using name.

- SolrQuery.addSortField got deprecated (use SolrQuery.addSort):
Introduced VersionUtil to determine if Solr 4.2.0 is in path by looking
for the changed ParseException. Still looking for a better way to do
this...

- Sorting on string fields did not work as expected:
Changed the test to sort on numeric field (popularity) worked fine, so
this seems to be a solr problem. Will have to check on this, but as
numeric fields work query translation seems to be fine so that the Issue
can be resolved.