DATASOLR-105 - Allow nesting of chained Criteria.
Chaining 'Criteria' did not allow combining nested criteria to span up a tree. Therefor it was not possible to build 'Queries' like: "field_1:foo OR (field_2:bar AND field_3:roo)" using the provided query API.

The implementation has changed under the hood to support this by building a tree of 'Nodes', in which 'Criteria' is a leave while 'Crotch' can be its parent, delegating addition of predicates to the most recent sibling.

The above sample can now be created as follows:
where("field_1").is("foo").or(where("field_2").is("bar").and("field_3").is("roo"));