DATASOLR-118 - MappingSolrConverter should honor generic type information within Maps and Collections.
Generic type information was not considered when reading values from Solr back into the entities in case of Collection and Map.

We now additionally check on the generic type of collections and maps and pipe the value retrieved from Solr through the conversion in case the type needs further treatment.