DATASOLR-147 - Upgrade to solr 4.8.
The basline for Spring Data projects is java 6. 

Lucene / Solr raised the baseline from java6 to java 7. As we have defined java 6 to be the baseline for Spring Data Projects we needed to tweak the maven compiler plugin a little in oder to run tests against the latest solr 4.8 version while keeping the source itself 1.6 compatible while running all the tests with java 1.7.

We did not experience any other glitches while running this, the only limitation we currently have is that the build has to be run under java 7 for a minimum.