DATASOLR-72 - Create schema for entities.
Initial setup introducing new types required to define schema for SolrPresitentEntity. 

SchemaWriter uses SolrSchemaRequest to read/write schema information from/to the solr JSON schema api. To make this work via SolrJ we’ve introduced according parsers using jackson for json conversion.

The schema api does not seem to work for embedded solr server - So we most likely will have to fall back to an http solr server instance for integration testsing, therfore we’d need to assume tests are only excuted when ther’s a running server available so that tests won’t fail when executed on CI server.

The lack of ability to remove fields from the schema once created, makes testing still a little harder - so it might be a good idea to rely on mocked tests.

Automatic core creation is another issue as the current CoreApi does not really allow to add new cores on the fly - at least not without prior interaction on the filesystem.

The CollectionAPI can also not be used for creating new cores it does not work properly in non cloud environments.