DATASOLR-170 - Add support for limiting the query result in the query derivation mechanism.
We check on the PartTree.isLimiting() and set an according PageRequest for the query. For Collection queries we now have to check for the eventual presence of a PageRequest in order to not override it with the max number of available documents.

In case a pageable is already provided for a limiting query we just take the limited amount of elements from that specific page, which means we’ve got to override the pageable to return an altered representation of offset and pageSize.