DATASOLR-163 - Register MappingContext properly for usage with spring-data-rest.
We now make sure we register a mapping context that is supposed to be used with sd-solr. In case one already defined a SimpleSolrMappingContext we’ll pick that one up for further usage.

This change allows the basic usage of sd-solr along with sd-rest which depends on reading mapping context.