DATASOLR-232 - Upgrade to Solr 5.
The upgrade requires a switch to SolrClient and its derivates. Since the inheritance changed quite a bit between SolrClient <> SolrServer (though those types are still available) in Version 5 supporting both types would have resulted in major drawbacks for the implementation.

Therefore this is a breaking change that demands a major version upgrade of spring-data-solr.

XML namspace and JavaConfig bean definition changes accordingly to reflect the changes.

  <solr:solr-client ../> replaces <solr:solr-server ../>

  @Bean solrClient() replaces @Bean solrServer()