DATASOLR-313 - Use native SolrJ Schema api.
We now use the Schema API provided by SolrJ (using the binary format) instead of the json based endpoints.

Also added a JUnit rule setting up an EmbeddedSolrServer that is able to deal with POST requests by guessing the request method based on the availability of ContentStreams in the request.