DATASOLR-324 - Fix SpellcheckOptions default request handler, dictionaries and extendedResults.
We removed the request handler defaulting to /spell for spellchecd queries. In case a dedicated endpoint is required the request handler can be set via the query itself.

Additionally added the missing extendedResults flag and changed the single dictionary option to multiple ones just as it is possible using vanilla solr configuration via xml.

And we added the extendedResults flag to both SolrOptions and @Spellcheck.